DatasetVision/requirements.txt
streamlit
torch
torchvision
transformers
pillow
numpy
pandas
matplotlib
plotly
scikit-image
opencv-python
timm
albumentations
huggingface-hub
tqdm
python-dotenv